initial refactor; not tested
